Hello World JFrame class
This is the very first Java GUI application I ever coded, this application is already done, nothing to change about it, that's why it went from "Master" branch to Hello-World branch. What I done to make this application is that I subclassted the JFrame class and used JFrame class's APIs. Well this project is done, moving on to next.