Confirm dialog with MD5 hash check
This project was fun, not only did I learned how to get the data from a dialog, but I also wrote MD5 hashing function which is used to generate a hash from a string. I also included to compare hashes and if they match, the application continues and if not it quits. The PIN is 28469.