Ditched next project, this is the final project
I was planning on working on the capturing the keystrokes, but I ditched this project and the learning OASD tag project is the last project, the final project goes to master branch. So I'm so thankful that you were following me through out this course, now I'm going to make a new repository for the Source Cipher. Thank you guys and girls :)